requirements.txt
tensorflow==2.20.0
numpy>=2.1.0
matplotlib==3.9.2
scikit-learn==1.5.2
seaborn==0.13.2
Pillow==10.4.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
jupyter==1.1.1
pytest==8.3.3
httpx==0.27.2
